requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
python-dotenv
tqdm